New ms

+++ lib/domain/models/veg_plan_detail_model.dart
@@ -124,7 +124,7 @@ class Schedules {
   int? invs;
   int? smbs;
   int? scheduledModules;
-  int? scheduledArea;
+  double? scheduledArea;
   String? plannedDate;
   int? status;
   String? statusShort;
